+++ frontend/src/modules/AdvancedCrudModule/DataTable.jsx
@@ -77,6 +77,7 @@ export default function DataTable({ config, extra = [] }) {
     },
   ];
 
+  const dispatch = useDispatch();
   const navigate = useNavigate();
 
   const handleRead = (record) => {
@@ -138,14 +139,14 @@ export default function DataTable({ config, extra = [] }) {
         >
           <EllipsisOutlined
             style={{ cursor: 'pointer', fontSize: '24px' }}
-            onClick={(e) => e.preventDefault()}
+            onClick={(e) => {
+              e.stopPropagation();
+            }}
           />
         </Dropdown>
       ),
     },
   ];
-
-  const dispatch = useDispatch();
 
   const handelDataTableLoad = (pagination) => {
     const options = { page: pagination.current || 1, items: pagination.pageSize || 10 };
@@ -188,6 +189,10 @@ export default function DataTable({ config, extra = [] }) {
         loading={listIsLoading}
         onChange={handelDataTableLoad}
         scroll={{ x: true }}
+        onRow={(record) => ({
+          onClick: () => handleRead(record),
+          style: { cursor: 'pointer' },
+        })}
       />
     </>
   );