+++ frontend/src/modules/ErpPanelModule/DataTable.jsx
@@ -79,6 +79,7 @@ export default function DataTable({ config, extra = [] }) {
     },
   ];
 
+  const dispatch = useDispatch();
   const navigate = useNavigate();
 
   const handleRead = (record) => {
@@ -141,14 +142,14 @@ export default function DataTable({ config, extra = [] }) {
         >
           <EllipsisOutlined
             style={{ cursor: 'pointer', fontSize: '24px' }}
-            onClick={(e) => e.preventDefault()}
+            onClick={(e) => {
+              e.stopPropagation();
+            }}
           />
         </Dropdown>
       ),
     },
   ];
-
-  const dispatch = useDispatch();
 
   const handelDataTableLoad = (pagination) => {
     const options = { page: pagination.current || 1, items: pagination.pageSize || 10 };
@@ -207,6 +208,10 @@ export default function DataTable({ config, extra = [] }) {
         loading={listIsLoading}
         onChange={handelDataTableLoad}
         scroll={{ x: true }}
+        onRow={(record) => ({
+          onClick: () => handleRead(record),
+          style: { cursor: 'pointer' },
+        })}
       />
     </>
   );